Added remove_from_ranges
