+++ encord/objects/ontology_labels_impl.py
@@ -1799,9 +1799,8 @@ class LabelRowV2:
         for frame in frames:
             present_frames.remove(frame)
 
-    def _remove_ranges_from_classification(self, classification: Classification, frames: Iterable[int]) -> None:
-        ranges = frames_to_ranges(list(frames))
-        self._classifications_to_ranges[classification].remove_ranges(ranges)
+    def _remove_ranges_from_classification(self, classification: Classification, ranges_to_remove: Ranges) -> None:
+        self._classifications_to_ranges[classification].remove_ranges(ranges_to_remove)
 
     def _add_to_frame_to_hashes_map(
         self, label_item: Union[ObjectInstance, ClassificationInstance], frames: Iterable[int]
