+++ tests/objects/test_label_structure.py
@@ -653,9 +653,9 @@ def test_add_and_get_classification_instances_to_audio_label_row(ontology):
     label_row = LabelRowV2(label_row_metadata, Mock(), ontology)
     label_row.from_labels_dict(EMPTY_AUDIO_LABELS)
 
-    classification_instance_1 = ClassificationInstance(text_classification)
-    classification_instance_2 = ClassificationInstance(text_classification)
-    classification_instance_3 = ClassificationInstance(checklist_classification)
+    classification_instance_1 = ClassificationInstance(text_classification, range_only=True)
+    classification_instance_2 = ClassificationInstance(text_classification, range_only=True)
+    classification_instance_3 = ClassificationInstance(checklist_classification, range_only=True)
 
     classification_instance_1.set_for_frames(Range(1, 2))
     classification_instance_2.set_for_frames(Range(3, 4))
@@ -675,7 +675,7 @@ def test_add_and_get_classification_instances_to_audio_label_row(ontology):
     filtered_classification_instances = label_row.get_classification_instances(text_classification)
     assert set(filtered_classification_instances) == {classification_instance_1, classification_instance_2}
 
-    overlapping_classification_instance = ClassificationInstance(text_classification)
+    overlapping_classification_instance = ClassificationInstance(text_classification, range_only=True)
     overlapping_classification_instance.set_for_frames(1)
 
     with pytest.raises(LabelRowError):
@@ -1045,3 +1045,91 @@ def test_classification_can_be_added_edited_and_removed(ontology):
 
     assert len(label_row.get_classification_instances()) == 0
 
+@pytest.fixture
+def ontology():
+    ontology_structure = PropertyMock(return_value=all_types_structure)
+    ontology = Mock(structure=ontology_structure)
+    yield ontology
+
+
+@pytest.fixture
+def empty_audio_label_row() -> LabelRowV2:
+    label_row_metadata_dict = asdict(FAKE_LABEL_ROW_METADATA)
+    label_row_metadata_dict["frames_per_second"] = 1000
+    label_row_metadata_dict["data_type"] = "AUDIO"
+    label_row_metadata = LabelRowMetadata(**label_row_metadata_dict)
+
+    label_row = LabelRowV2(label_row_metadata, Mock(), ontology_from_dict(all_ontology_types))
+    label_row.from_labels_dict(EMPTY_AUDIO_LABELS)
+
+    return label_row
+
+
+def test_non_range_classification_cannot_be_added_to_audio_label_row(ontology):
+    label_row_metadata_dict = asdict(FAKE_LABEL_ROW_METADATA)
+    label_row_metadata_dict["frames_per_second"] = 1000
+    label_row_metadata_dict["data_type"] = "AUDIO"
+    label_row_metadata = LabelRowMetadata(**label_row_metadata_dict)
+
+    label_row = LabelRowV2(label_row_metadata, Mock(), ontology_from_dict(all_ontology_types))
+    label_row.from_labels_dict(EMPTY_AUDIO_LABELS)
+
+    with pytest.raises(LabelRowError):
+        classification_instance = ClassificationInstance(checklist_classification)
+        classification_instance.set_for_frames(Range(start=0, end=1500))
+        label_row.add_classification_instance(classification_instance)
+
+    with pytest.raises(LabelRowError):
+        classification_instance = checklist_classification.create_instance()
+        label_row.add_classification_instance(classification_instance)
+
+
+def test_audio_classification_overwrite(ontology, empty_audio_label_row: LabelRowV2):
+    classification_instance = ClassificationInstance(checklist_classification, range_only=True)
+    classification_instance.set_for_frames(Range(start=0, end=100))
+    empty_audio_label_row.add_classification_instance(classification_instance)
+
+    with pytest.raises(LabelRowError):
+        classification_instance.set_for_frames(Range(start=5, end=20))
+
+    with pytest.raises(LabelRowError):
+        classification_instance.set_for_frames(Range(start=100, end=101))
+
+    # No error when set overwrite to True
+    classification_instance.set_for_frames(Range(start=100, end=101), overwrite=True)
+    range_list = classification_instance.range_list
+    assert len(range_list) == 1
+    assert range_list[0].start == 0
+    assert range_list[0].end == 101
+
+
+def test_audio_classification_exceed_max_frames(ontology, empty_audio_label_row: LabelRowV2):
+    classification_instance = ClassificationInstance(checklist_classification, range_only=True)
+    classification_instance.set_for_frames(Range(start=0, end=100))
+    empty_audio_label_row.add_classification_instance(classification_instance)
+
+    with pytest.raises(LabelRowError):
+        classification_instance.set_for_frames(Range(start=200, end=5000))
+
+
+def test_audio_classification_can_be_added_edited_and_removed(ontology, empty_audio_label_row: LabelRowV2):
+    label_row = empty_audio_label_row
+    classification_instance = ClassificationInstance(checklist_classification, range_only=True)
+    classification_instance.set_for_frames(Range(start=0, end=1500))
+    range_list = classification_instance.range_list
+    assert len(range_list) == 1
+    assert range_list[0].start == 0
+    assert range_list[0].end == 1500
+
+    label_row.add_classification_instance(classification_instance)
+    assert len(label_row.get_classification_instances()) == 1
+    classification_instance.set_for_frames(Range(start=2000, end=2499))
+    range_list = classification_instance.range_list
+    assert len(range_list) == 2
+    assert range_list[0].start == 0
+    assert range_list[0].end == 1500
+    assert range_list[1].start == 2000
+    assert range_list[1].end == 2499
+
+    label_row.remove_classification(classification_instance)
+    assert len(label_row.get_classification_instances()) == 0
